normalize keys in order to use assq
as explained in this blog post [1], the `assq` function has its own op
code, which makes it miles faster than `assoc` and similar. for that
reason, it is best if we normalize the keys to either strings or
symbols (in the case of this function, given that we need to check if it
is not nil and to keep it backwards compatible, both) and work with
that.

[1] https://nullprogram.com/blog/2017/01/30/