allow verb-map-request functions to be anonymous
instead of using only symbols, when we call (read arg) on a string it
gets read as a lisp expression and returns a lisp object. if the string
is just some characters, such as "foo", "bar", "some-name", the
resulting expression is a symbol, just like the previous behaviour.

but this change also allows us to provied the value of verb-map-request
as a lambda expression, which gets correctly read as such and returned
as a list whose car is 'lambda (precisely what a lambda expression is,
in the end of things). hence why we now check if the resulting fn is a
function with functionp instead of relying only on named functions.