ensure the same property isn't duplicated
with the previous implementation a property that span more than one
line, using the :property+: syntax, but used different cases for each of
those lines, e.g.

:Property: foo
:property+: bar

these multi-line properties were getting duplicated in the request spec.

now, we already create the property list upcased, ensuring that the same
property with different case won't be added to the list if it is already
present.