refactor: Extracted three functions out from the METHOD+URL section
This step is necessary as the plan is to turn the METHOD+URL section
into METHOD+URL+PROTOCOL, as suggested in the original issue: "get
http://example.com http/1.1". More code will obviously have to be
added in order match protocols so in this step I tried to simplify the
METHOD+URL section and extract out functions that could also be used
to match protocols. Tests were are added for all three functions that
I extracted out.