refactor: METHOD+URL+PROTOCOL section has been added in place of the METHOD+URL section in verb-request-spec-from-string
The refactor started in my last commit is now finished. It would have
been very difficult to add protocol to the original METHOD+URL
section as it already contained nested conditional logic. Instead, I
wrote a verb--single-line-method-url-protocol, a function that splits
the method+url+protocol line in to parts. Method and protocol are then
validated using validation functions that I added for this commit while the
url is simply stored. If the url is on multiple lines, the lines are
concatenated together and then if a protocol is also specified, it's
stored. Once I tested this refactor, I added protocol to the
verb-request-spec created at the the of verb-request-spec-from-string function.